[WIP] Indexer utility for compressing ASTs
Next step is to compute the bit mask through a trie
of the indexed paths.

Example:

has(a.b.c) && has(a.b.c.d) ? x : y

The computation of a.b.c presence should not be revisited,
but rather should feed into the next index, so in that sense
a trie that preseves the branches, but resets visit presence
based on last observed value would be relatively cheap.